% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_utilities.R
\name{are_models_nested}
\alias{are_models_nested}
\title{Check if Two Models are Nested}
\usage{
are_models_nested(model1, model2)
}
\arguments{
\item{model1}{An object of class \code{lm}, \code{glm}, or another model
type with a \code{coef} method. Represents the first model to be
compared.}

\item{model2}{An object of class \code{lm}, \code{glm}, or another model
type with a \code{coef} method. Represents the second model to be
compared.}
}
\value{
A logical value: \code{TRUE} if one model is nested within the other
(i.e., all the terms in one model are contained within the other),
and \code{FALSE} otherwise.
}
\description{
This function determines whether one model is nested within another. A model
is considered nested if all of its terms are a subset of the terms in the
other model.
}
\examples{
# Example usage with linear models
model1 <- lm(mpg ~ wt, data = mtcars)
model2 <- lm(mpg ~ wt + hp, data = mtcars)
are_models_nested(model1, model2) # TRUE, model1 is nested within model2

model3 <- lm(mpg ~ qsec, data = mtcars)
are_models_nested(model1, model3) # FALSE, no nesting between model1 and model3

}
