% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{coef.mvreg}
\alias{coef.mvreg}
\title{Extract mvreg coefficients}
\usage{
\method{coef}{mvreg}(object, partition = c("all", "mu", "s2"), ...)
}
\arguments{
\item{object}{A mvreg object.}

\item{partition}{A character vector that specifies which partition of coefficients to return.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Specified coefficients of mvreg model.
}
\description{
Extract mvreg coefficients
}
\examples{
mvreg_mod <- mvreg(Sepal.Length ~ Species, ~ Species, data = iris)
summary(mvreg_mod)
coef(mvreg_mod)
coef(mvreg_mod, partition = "s2")
}
