% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{d2l}
\alias{d2l}
\title{Full hessian of model loglikelihood}
\usage{
d2l(y, x, z, b, t)
}
\arguments{
\item{y}{Vector of response variable.}

\item{x}{Matrix of explanatory variables for mean component.}

\item{z}{Matrix of explanatory variables for variance component.}

\item{b}{Vector of parameters for mean component.}

\item{t}{Vector of parameters for variance component.}
}
\value{
Full hessian of model loglikelihood.
}
\description{
Full hessian of model loglikelihood
}
