% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{d2ldb}
\alias{d2ldb}
\title{Hessian of model loglikelihood with respect to beta}
\usage{
d2ldb(y, x, z, b, t)
}
\arguments{
\item{y}{Vector of response variable.}

\item{x}{Matrix of explanatory variables for mean component.}

\item{z}{Matrix of explanatory variables for variance component.}

\item{b}{Vector of parameters for mean component.}

\item{t}{Vector of parameters for variance component.}
}
\value{
Hessian of loglikelihood with respect to beta.
}
\description{
Hessian of model loglikelihood with respect to beta
}
