% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{d2ldt}
\alias{d2ldt}
\title{Hessian of model loglikelihood with respect to tau}
\usage{
d2ldt(y, x, z, b, t)
}
\arguments{
\item{y}{Vector of response variable.}

\item{x}{Matrix of explanatory variables for mean component.}

\item{z}{Matrix of explanatory variables for variance component.}

\item{b}{Vector of parameters for mean component.}

\item{t}{Vector of parameters for variance component.}
}
\value{
Hessian of loglikelihood with respect to tau.
}
\description{
Hessian of model loglikelihood with respect to tau
}
