% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{dldt}
\alias{dldt}
\title{Gradient of model loglikelihood with respect to tau}
\usage{
dldt(y, x, z, b, t)
}
\arguments{
\item{y}{Vector of response variable.}

\item{x}{Matrix of explanatory variables for mean component.}

\item{z}{Matrix of explanatory variables for variance component.}

\item{b}{Vector of parameters for mean component.}

\item{t}{Vector of parameters for variance component.}
}
\value{
Gradient with respect to tau.
}
\description{
Gradient of model loglikelihood with respect to tau
}
