% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{fitted.mvreg}
\alias{fitted.mvreg}
\title{Fitted values of mvreg model}
\usage{
\method{fitted}{mvreg}(x, type = c("all", "mu", "log.s2", "s2"))
}
\arguments{
\item{x}{A mvreg object.}

\item{type}{A character vector specifying the component for which to return fitted values.}
}
\value{
Fitted values for a given component of a mvreg model.
}
\description{
Fitted values of mvreg model
}
\examples{
mvreg_mod <- mvreg(Sepal.Length ~ Species, ~Sepal.Width, data = iris) # different formulas
# fitted
fitted(mvreg_mod)
fitted(mvreg_mod, type = "mu")
fitted(mvreg_mod, type = "log.s2")
fitted(mvreg_mod, type = "s2")
}
