% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{ll}
\alias{ll}
\title{Loglikelihood of heteroskedastic linear model}
\usage{
ll(y, x, z, b, t)
}
\arguments{
\item{y}{Vector of response variable.}

\item{x}{Matrix of explanatory variables for mean component.}

\item{z}{Matrix of explanatory variables for variance component.}

\item{b}{Vector of parameters for mean component.}

\item{t}{Vector of parameters for variance component.}
}
\value{
Value of loglikelihood for a given sample and parameters' values.
}
\description{
Loglikelihood of heteroskedastic linear model
}
