% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{logLik.mvreg}
\alias{logLik.mvreg}
\title{logLik method for mvreg}
\usage{
\method{logLik}{mvreg}(object, ...)
}
\arguments{
\item{object}{A \code{mvreg} object.}

\item{...}{Additional optional arguments (not currently used).}
}
\value{
A log-likelihood value of the fitted \code{mvreg} model, with attributes:
\item{df}{Degrees of freedom associated with the model.}
\item{nobs}{Number of observations used in the fitting.}
}
\description{
logLik method for mvreg
}
\examples{
mod <- mvreg(Sepal.Length ~ Species, data = iris)
logLik(mod) # returns the log-likelihood of the fitted model
}
