% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_fit.R
\name{mvreg.fit}
\alias{mvreg.fit}
\title{Fitter function for mvreg}
\usage{
mvreg.fit(y, x, z, b0, t0, tol = 1e-10, maxit = 100)
}
\arguments{
\item{y}{Response variable vector.}

\item{x}{Design matrix for mean component.}

\item{z}{Design matrix for variance component.}

\item{b0}{Starting values for parameters of mean component.}

\item{t0}{Starting values for parameters of variance component.}

\item{tol}{Positive value indicating what is the minimum difference between parameter estimates between two iterations to stop the algorithm.}

\item{maxit}{Integer value indicating the maximum number of iteration.}
}
\value{
Estimate of coefficients and variance-covariance matrix
}
\description{
Fitter function for mvreg
}
