% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_start.R
\name{mvreg.start}
\alias{mvreg.start}
\title{Starting values for mvreg}
\usage{
mvreg.start(y, x, z, start.s2 = c("residuals", "gamma", "zero"))
}
\arguments{
\item{y}{Response variable vector.}

\item{x}{Design matrix for mean component.}

\item{z}{Design matrix for variance component.}

\item{start.s2}{A character vector indicating how to select initial values for variance component parameters.}
}
\value{
Starting values of estimates of parameters
}
\description{
Starting values for mvreg
}
