% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_fit.R
\name{mvreg_fit}
\alias{mvreg_fit}
\title{Fitter function for mvreg}
\usage{
mvreg_fit(
  y,
  x,
  z,
  b0,
  t0,
  tol = 0.0000000001,
  maxit = 100,
  method = c("wls", "full_nr"),
  vcov.type = c("expected", "observed")
)
}
\arguments{
\item{y}{Response variable vector.}

\item{x}{Design matrix for mean component.}

\item{z}{Design matrix for variance component.}

\item{b0}{Starting values for parameters of mean component.}

\item{t0}{Starting values for parameters of variance component.}

\item{tol}{Positive value indicating what is the minimum difference between parameter estimates between two iterations to stop the algorithm.}

\item{maxit}{Integer value indicating the maximum number of iteration.}

\item{method}{Method chosen for estimation of parameters of mean component}

\item{vcov.type}{A string to specify whether to use observed or expected Fisher information matrix in order to compute variance-covariance matrix of estimates}
}
\value{
Estimate of coefficients and variance-covariance matrix
}
\description{
Fitter function for mvreg
}
\examples{
n <- 100
x1 <- rnorm(n)
x2 <- factor(sample(letters[1:3], n, TRUE))
x <- model.matrix(~ x1 + x2)
z1 <- factor(sample(letters[1:3], n, TRUE))
z <- model.matrix(~z1)

b <- rnorm(ncol(x))
t <- rnorm(ncol(z))

y <- rnorm(n, mean = x \%*\% b, sd = sqrt(exp(z \%*\% t)))

start.list <- mvreg_start(y, x, z, start.s2 = "gamma")
b0 <- start.list$b0
t0 <- start.list$t0

mvreg_fit(y, x, z, b0, t0)
}
