% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{mvreg_gradient_s2}
\alias{mvreg_gradient_s2}
\title{Gradient of model log-likelihood with respect to variance coefficients (tau)}
\usage{
mvreg_gradient_s2(y, x, z, b, t)
}
\arguments{
\item{y}{A numeric vector representing the response variable.}

\item{x}{A numeric matrix of explanatory variables for the mean component of the model.}

\item{z}{A numeric matrix of explanatory variables for the variance component of the model.}

\item{b}{A numeric vector of parameters for the mean component.}

\item{t}{A numeric vector of parameters for the variance component.}
}
\value{
A numeric vector representing the gradient of the log-likelihood with respect to the variance component coefficients \eqn{\boldsymbol{\tau}}.
}
\description{
This function computes the gradient of the log-likelihood of a heteroskedastic linear model with respect to the coefficients of the variance component (denoted as \eqn{\boldsymbol{\tau}}).
}
\details{
The gradient is calculated based on the partial derivatives of the log-likelihood function with respect to the variance coefficients \eqn{\boldsymbol{\tau}}.
The \eqn{j}-th element of the gradient is computed as:
\deqn{\dfrac{\partial \ell}{\partial \tau_j} = -\dfrac{1}{2}\sum_{i=1}^n\left\{ 1 - \dfrac{\left(y_i - \mathbf{x}_i'\boldsymbol{\beta}\right)^2}{\exp\left\{\mathbf{z}_i'\boldsymbol{\tau}\right\}}\right\}z_{ij}}
}
\examples{
# Fit a heteroskedastic linear model to the iris dataset
mvreg_mod <- mvreg(Sepal.Length ~ Species, data = iris) # same formula for mean and variance

# Extract model components
y <- mvreg_mod$y # Response variable
x <- mvreg_mod$x # Model matrix for the mean component
z <- mvreg_mod$z # Model matrix for the variance component
b <- coef(mvreg_mod, "mu") # Coefficients of the mean component
t <- coef(mvreg_mod, "s2") # Coefficients of the variance component

# Compute gradient with respect to the variance coefficients
mvreg_gradient_s2(y, x, z, b, t)
}
