% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{mvreg_hessian}
\alias{mvreg_hessian}
\title{Full hessian of model loglikelihood}
\usage{
mvreg_hessian(y, x, z, b, t, type = c("observed", "expected"))
}
\arguments{
\item{y}{Vector of response variable.}

\item{x}{Matrix of explanatory variables for mean component.}

\item{z}{Matrix of explanatory variables for variance component.}

\item{b}{Vector of parameters for mean component.}

\item{t}{Vector of parameters for variance component.}

\item{type}{A string to specify whether to output the observed or expected matrix}
}
\value{
Full hessian of model loglikelihood.
}
\description{
Full hessian of model loglikelihood
}
\examples{
mvreg_mod <- mvreg(Sepal.Length ~ Species, data = iris) # same formula for mean and variance
y <- mvreg_mod$y # response variable
x <- mvreg_mod$x # model.matrix for mean component
z <- mvreg_mod$z # model.matrix for variance component
b <- coef(mvreg_mod, "mu") # coefficients of mean component
t <- coef(mvreg_mod, "s2") # coefficients of variance component

# second derivatives
mvreg_hessian_mu(y, x, z, b, t) # w.r.t. mean coefficients
mvreg_hessian_s2(y, x, z, b, t) # w.r.t. variance coefficients
mvreg_hessian_mus2(y, x, z, b, t) # w.r.t. mean coefficients and variance coefficients
mvreg_hessian(y, x, z, b, t) # full hessian

}
