% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{mvreg_hessian_mu}
\alias{mvreg_hessian_mu}
\title{Hessian of model log-likelihood with Respect to mean coefficients (beta)}
\usage{
mvreg_hessian_mu(y, x, z, b, t, type = c("observed", "expected"))
}
\arguments{
\item{y}{A numeric vector representing the response variable.}

\item{x}{A numeric matrix of explanatory variables for the mean component of the model.}

\item{z}{A numeric matrix of explanatory variables for the variance component of the model.}

\item{b}{A numeric vector of parameters for the mean component.}

\item{t}{A numeric vector of parameters for the variance component.}

\item{type}{A character string specifying whether to return the observed or expected Hessian matrix.
The default is "observed".}
}
\value{
A numeric matrix representing the Hessian of the log-likelihood with respect to the mean component coefficients \eqn{\boldsymbol{\beta}}.
}
\description{
This function computes the Hessian matrix of the log-likelihood of a heteroskedastic linear model with respect to the coefficients of the mean component.
The Hessian can be calculated as either the observed or expected matrix, depending on the \code{type} argument.
}
\details{
The Hessian is calculated as the matrix of second-order partial derivatives of the log-likelihood function with respect to \eqn{\boldsymbol{\beta}}.
The \eqn{(j, l)}-th element of the Hessian is computed using the formula:
\deqn{\dfrac{\partial^2 \ell}{\partial \beta_j \partial \beta_l} = -\sum_{i=1}^n \dfrac{x_{ij}x_{il}}{\exp\left\{\mathbf{z}_i'\boldsymbol{\tau}\right\}}}
In this case, the observed hessian and the expected hessian coincide. The argument \code{type} is maintained just for consistency.
}
\examples{
# Fit a heteroskedastic linear model to the iris dataset
mvreg_mod <- mvreg(Sepal.Length ~ Species, data = iris) # same formula for mean and variance

# Extract model components
y <- mvreg_mod$y        # Response variable
x <- mvreg_mod$x        # Model matrix for the mean component
z <- mvreg_mod$z        # Model matrix for the variance component
b <- coef(mvreg_mod, "mu")  # Coefficients of the mean component
t <- coef(mvreg_mod, "s2")  # Coefficients of the variance component

# Compute Hessian with respect to the mean coefficients
mvreg_hessian_mu(y, x, z, b, t, type = "observed")
mvreg_hessian_mu(y, x, z, b, t, type = "expected")
}
