% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{mvreg_hessian_mus2}
\alias{mvreg_hessian_mus2}
\title{Hessian of model log-likelihood with respect to mean and variance coefficients (beta and tau)}
\usage{
mvreg_hessian_mus2(y, x, z, b, t, type = c("observed", "expected"))
}
\arguments{
\item{y}{A numeric vector representing the response variable.}

\item{x}{A numeric matrix of explanatory variables for the mean component of the model.}

\item{z}{A numeric matrix of explanatory variables for the variance component of the model.}

\item{b}{A numeric vector of parameters for the mean component.}

\item{t}{A numeric vector of parameters for the variance component.}

\item{type}{A character string specifying whether to return the observed or expected Hessian matrix.
The default is "observed".}
}
\value{
A numeric matrix representing the Hessian of the log-likelihood with respect to the mean and variance component coefficients.
}
\description{
This function computes the Hessian matrix of the log-likelihood of a heteroskedastic linear model
with respect to the coefficients of the mean component (denoted as \eqn{\boldsymbol{\beta}})
and the coefficients of the variance component (denoted as \eqn{\boldsymbol{\tau}}).
The Hessian can be calculated as either the observed or expected matrix, depending on the \code{type} argument.
}
\details{
The observed Hessian is calculated as the matrix of mixed second-order partial derivatives of the log-likelihood function
with respect to \eqn{\boldsymbol{\beta}} and \eqn{\boldsymbol{\tau}}.
The \eqn{(j, l)}-th element of the observed Hessian is computed using the formula:
\deqn{\dfrac{\partial^2 \ell}{\partial \beta_j \partial \tau_j} = -\sum_{i=1}^n \dfrac{y_i - \mathbf{x}_i'\boldsymbol{\beta}}{\exp\left\{\mathbf{z}_i'\boldsymbol{\tau}\right\}}x_{ij}z_{il}}

The expected Hessian is a zero matrix.
}
\examples{
# Fit a heteroskedastic linear model to the iris dataset
mvreg_mod <- mvreg(Sepal.Length ~ Species, data = iris) # same formula for mean and variance

# Extract model components
y <- mvreg_mod$y # Response variable
x <- mvreg_mod$x # Model matrix for the mean component
z <- mvreg_mod$z # Model matrix for the variance component
b <- coef(mvreg_mod, "mu") # Coefficients of the mean component
t <- coef(mvreg_mod, "s2") # Coefficients of the variance component

# Compute Hessian with respect to mean and variance coefficients
mvreg_hessian_mus2(y, x, z, b, t, type = "observed")
mvreg_hessian_mus2(y, x, z, b, t, type = "expected")
}
