% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{mvreg_hessian_s2}
\alias{mvreg_hessian_s2}
\title{Hessian of model log-likelihood with respect to variance coefficients (tau)}
\usage{
mvreg_hessian_s2(y, x, z, b, t, type = c("observed", "expected"))
}
\arguments{
\item{y}{A numeric vector representing the response variable.}

\item{x}{A numeric matrix of explanatory variables for the mean component of the model.}

\item{z}{A numeric matrix of explanatory variables for the variance component of the model.}

\item{b}{A numeric vector of parameters for the mean component.}

\item{t}{A numeric vector of parameters for the variance component.}

\item{type}{A character string specifying whether to return the observed or expected Hessian matrix.
The default is "observed".}
}
\value{
A numeric matrix representing the Hessian of the log-likelihood with respect to the variance component coefficients \eqn{\boldsymbol{\tau}}.
}
\description{
This function computes the Hessian matrix of the log-likelihood of a heteroskedastic linear model with respect to the coefficients of the variance component.
The Hessian can be calculated as either the observed or expected matrix, depending on the \code{type} argument.
}
\details{
The observed Hessian is calculated as the matrix of second-order partial derivatives of the log-likelihood function with respect to \eqn{\boldsymbol{\tau}}.
The \eqn{(j, l)}-th element of the observed Hessian is computed using the formula:
\deqn{\dfrac{\partial^2 \ell}{\partial \tau_j \partial \tau_l} = -\dfrac{1}{2}\sum_{i=1}^n \dfrac{\left(y_i - \mathbf{x}_i'\boldsymbol{\beta}\right)^2}{\exp\left\{\mathbf{z}_i'\boldsymbol{\tau}\right\}}z_{ij}z_{il} }

The \eqn{(j, l)}-th element of the expected Hessian is computed calculating the expectation of the \eqn{(j, l)}-th element of the observed Hessian:
\deqn{\mathbb{E}\left(\dfrac{\partial^2 \ell}{\partial \tau_j \partial \tau_l}\right) = -\dfrac{1}{2}\sum_{i=1}^n z_{ij}z_{il}}
}
\examples{
# Fit a heteroskedastic linear model to the iris dataset
mvreg_mod <- mvreg(Sepal.Length ~ Species, data = iris) # same formula for mean and variance

# Extract model components
y <- mvreg_mod$y # Response variable
x <- mvreg_mod$x # Model matrix for the mean component
z <- mvreg_mod$z # Model matrix for the variance component
b <- coef(mvreg_mod, "mu") # Coefficients of the mean component
t <- coef(mvreg_mod, "s2") # Coefficients of the variance component

# Compute Hessian with respect to the variance coefficients
mvreg_hessian_s2(y, x, z, b, t, type = "observed")
mvreg_hessian_s2(y, x, z, b, t, type = "expected")
}
