% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_simul.R
\name{mvreg_simul}
\alias{mvreg_simul}
\title{Simulation Study for Multivariate Regression Models}
\usage{
mvreg_simul(
  x,
  z,
  b,
  t,
  nsim = 100,
  sig.level = 0.05,
  seed = NULL,
  method = c("wls", "full_nr"),
  vcov.type = c("expected", "observed"),
  start.s2 = c("residuals", "gamma", "zero"),
  tol = 1e-10,
  maxit = 100
)
}
\arguments{
\item{x}{A numeric matrix representing the design matrix for the mean component.
Each column corresponds to a predictor variable. The first column should be
an intercept term.}

\item{z}{A numeric matrix representing the design matrix for the variance component.
Each column corresponds to a predictor variable for the variance. The first
column should also be an intercept term.}

\item{b}{A numeric vector of true values for the parameters of the mean component.
Its length must match the number of columns in \code{x}.}

\item{t}{A numeric vector of true values for the parameters of the variance component.
Its length must match the number of columns in \code{z}.}

\item{nsim}{An integer indicating the number of samples to simulate. Default is \code{100}.}

\item{sig.level}{A numeric value representing the significance level for Wald tests.
Default is \code{0.05}.}

\item{seed}{An optional integer to set the seed for random number generation. This ensures
reproducibility of the results. Default is NULL.}

\item{method}{A character vector indicating the method chosen for estimation of parameters of mean component. See the documentation of \code{\link{mvreg_fit}} for details.}

\item{vcov.type}{A character vector to specify whether to use observed or expected Fisher information matrix in order to compute variance-covariance matrix of estimates. See the documentation of \code{\link{mvreg_fit}}, \code{\link{mvreg_hessian_mu}}, \code{\link{mvreg_hessian_s2}}, \code{\link{mvreg_hessian_mus2}} and \code{\link{mvreg_hessian}} for details.}

\item{start.s2}{A character vector indicating how to select initial values for variance component parameters. See the documentation of \code{\link{mvreg_start}} for details.}

\item{tol}{A positive numeric value indicating the minimum difference between parameter
estimates between two iterations to stop the algorithm. Default is \code{1e-10}.}

\item{maxit}{An integer value specifying the maximum number of iterations for the fitting
algorithm. Default is \code{100}.}
}
\value{
A list containing results of the simulation study, including:
\describe{
\item{tab}{A data frame summarizing the true values, mean estimates, distortion,
variance, standard errors (SE), mean squared error (MSE), root mean squared
error (RMSE), and proportion of rejected null hypotheses.}
\item{theta}{A data frame of parameter estimates from the simulated models.}
\item{mean_vtheta}{A matrix containing the average variance-covariance matrices of the
estimates across simulations.}
\item{vtheta}{A list of variance-covariance matrices for each simulated model.}
\item{it}{A vector of the number of iterations taken to converge for each simulation.}
\item{y}{A data frame of the simulated response variables.}
\item{n}{The number of observations used in the simulation.}
\item{nsim}{The number of simulations performed.}
\item{seed}{The seed used for random number generation.}
\item{k}{The number of predictors in the mean component.}
\item{p}{The number of predictors in the variance component.}
\item{converged}{The number of simulations that converged successfully.}
\item{non_converged}{The number of simulations that did not converge.}
\item{total_time}{The total time taken to perform the simulations.}
}
}
\description{
This function performs a simulation study to evaluate the properties of
estimators for the parameters of \code{mvreg} models. It generates
simulated response variables based on the specified design matrices for the
mean and variance components and true parameter values, and then fits the
model to the simulated data.
}
\examples{
n <- 100
x <- cbind(1, rnorm(n)) # Design matrix for mean component
z <- cbind(1, x[, 2], x[, 2]^2) # Design matrix for variance component
b <- rnorm(ncol(x)) # True parameters for mean component
t <- rnorm(ncol(z)) # True parameters for variance component
mvreg_simul(x, z, b, t, nsim = 100, seed = 43)
}
