% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_start.R
\name{mvreg_start}
\alias{mvreg_start}
\title{Starting values for mvreg}
\usage{
mvreg_start(y, x, z, start.s2 = c("residuals", "gamma", "zero"))
}
\arguments{
\item{y}{Response variable vector.}

\item{x}{Design matrix for mean component.}

\item{z}{Design matrix for variance component.}

\item{start.s2}{A character vector indicating how to select initial values for variance component parameters.}
}
\value{
Starting values of estimates of parameters
}
\description{
Starting values for mvreg
}
\examples{
n <- 100
x1 <- rnorm(n)
x2 <- factor(sample(letters[1:3], n, TRUE))
x <- model.matrix(~ x1 + x2)
z1 <- factor(sample(letters[1:3], n, TRUE))
z <- model.matrix(~z1)

b <- rnorm(ncol(x))
t <- rnorm(ncol(z))

y <- rnorm(n, mean = x \%*\% b, sd = sqrt(exp(z \%*\% t)))

mvreg_start(y, x, z, start.s2 = "residuals")
mvreg_start(y, x, z, start.s2 = "gamma")
mvreg_start(y, x, z, start.s2 = "zero")
}
