% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{predict.mvreg}
\alias{predict.mvreg}
\title{Predict method for mvreg}
\usage{
\method{predict}{mvreg}(
  object,
  type = c("all", "mu", "log.s2", "s2"),
  newdata,
  se.fit = F,
  interval = F,
  sig.level = 0.95
)
}
\arguments{
\item{object}{A mvreg object.}

\item{type}{A character vector specifying the component for which to return predicted values.}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{se.fit}{A logical value indicating if standard errors are to be returned.}

\item{interval}{A logical value indicating if confidence intervals are to be returned.}

\item{sig.level}{Confidence level for confidence intervals.}
}
\description{
Predict method for mvreg
}
