% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_anova.R
\name{print.anova.mvreg}
\alias{print.anova.mvreg}
\title{Print Method for ANOVA Results of mvreg models}
\usage{
\method{print}{anova.mvreg}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{anova.mvreg}, which contains the
results of the ANOVA analysis.}

\item{digits}{An integer specifying the number of significant digits
to print. Default is the maximum of 3 or the \code{digits} option minus 3.}

\item{...}{Additional arguments (not used).}
}
\value{
This function prints the results directly to the console and does not
return a value.
}
\description{
This function formats and prints the results of an ANOVA-like analysis
for \code{mvreg} objects. It provides a summary of the tests conducted on
the mean and variance components of the models.
}
\details{
If only one model was provided to the \code{anova.mvreg} function, the output
includes tests for both the mean and variance components, along with their
respective formulas. If multiple models were provided, it summarizes the
likelihood ratio tests for the models compared, along with their formulas.
}
\examples{
# Fit two models
mod1 <- mvreg(Sepal.Length ~ Species + Sepal.Width, data = iris)
mod2 <- mvreg(Sepal.Length ~ Species, ~1, data = iris)
anova(mod1, mod2)

# Fit a single model
mod3 <- mvreg(Sepal.Length ~ Species, ~ Sepal.Width, data = iris)
anova(mod3)
}
