% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{print.summary.mvreg}
\alias{print.summary.mvreg}
\title{Print method for summary.mvreg objects}
\usage{
\method{print}{summary.mvreg}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{A \code{summary.mvreg} object.}

\item{digits}{Minimal number of significant digits, see \code{print.default}.}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A printed summary of the \code{mvreg} model, including:
\item{Call}{The call used to fit the model.}
\item{Residuals}{Summary statistics of the residuals from the fitted model.}
\item{Coefficients}{Coefficients for the mean and log(variance) components, including standard errors and significance.}
\item{Model Fit Statistics}{Log-likelihood, AIC, and BIC of the fitted model.}
\item{Likelihood Ratio Test (LRT)}{Comparison with a linear model (lm), including likelihood ratio, degrees of freedom, and p-value.}
}
\description{
Print method for summary.mvreg objects
}
\examples{
mod <- mvreg(Sepal.Length ~ Species, data = iris)
s <- summary(mod)
print(s)
}
