% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{summary.mvreg}
\alias{summary.mvreg}
\title{Summary method for mvreg}
\usage{
\method{summary}{mvreg}(object, ...)
}
\arguments{
\item{object}{A mvreg object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A summary.mvreg object.
}
\description{
Summary method for mvreg
}
