% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_anova.R
\name{anova.mvreg}
\alias{anova.mvreg}
\title{ANOVA for mvreg models}
\usage{
\method{anova}{mvreg}(object, ...)
}
\arguments{
\item{object}{A \code{mvreg} object. If only one \code{mvreg} object is specified, the model is compared with all the reduced models obtained by sequentially dropping one term at a time from the original formula.}

\item{...}{Additional \code{mvreg} objects to be compared with the first model.}
}
\value{
A list of class \code{anova.mvreg} containing:
\itemize{
\item If one model is provided (i.e., \code{...} is empty):
\itemize{
\item \code{mu.tests}: A data frame with results of the ANOVA for the mean component, including:
\itemize{
\item \verb{-2logLik}: The negative two times the log-likelihood of each model.
\item \code{n.param}: The number of parameters in each model.
\item \code{LRT}: The likelihood ratio test statistic.
\item \code{df}: The degrees of freedom for the test.
\item \verb{Pr(>Chi)}: The p-value associated with the likelihood ratio test.
}
\item \code{s2.tests}: A similar data frame for the variance component.
\item \code{new.formula.mu}: A character vector of formulas used for the mean component.
\item \code{new.formula.s2}: A character vector of formulas used for the variance component.
\item \code{n_models}: The number of models compared.
}
\item If more than one model is provided:
\itemize{
\item \code{tests}: A data frame with results of the likelihood ratio tests, including:
\itemize{
\item \verb{-2logLik}: The negative two times the log-likelihood of each model.
\item \code{n.param}: The number of parameters in each model.
\item \code{LRT}: The likelihood ratio test statistic.
\item \verb{Pr(>Chi)}: The p-value associated with the likelihood ratio test.
}
\item \code{formulas.mu}: A character vector of formulas used for the mean component.
\item \code{formulas.s2}: A character vector of formulas used for the variance component.
\item \code{n_models}: The number of models compared.
}
}
}
\description{
This function performs an ANOVA-like analysis for \code{mvreg} objects to compare models
with different formulas for the mean and variance components. It calculates the likelihood ratio test
statistics and their associated p-values.
}
\examples{
# Fit two models
mod1 <- mvreg(Sepal.Length ~ Species + Sepal.Width, data = iris)
mod2 <- mvreg(Sepal.Length ~ Species, ~1, data = iris)
anova(mod1, mod2)

# Fit a single model
mod3 <- mvreg(Sepal.Length ~ Species, ~ Sepal.Width, data = iris)
anova(mod3)
}
