% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_anova.R
\name{anova.mvreg}
\alias{anova.mvreg}
\title{ANOVA-like Comparison for mvreg Models}
\usage{
\method{anova}{mvreg}(object, ..., order.models = T)
}
\arguments{
\item{object}{A \code{mvreg} object, representing the main model to compare.
If this is the only model provided, the function compares this model with reduced versions of itself.}

\item{...}{Additional \code{mvreg} objects to compare with the first model.}

\item{order.models}{A logical value indicating if the models should be ordered.
If \code{TRUE} (default), models are ranked by the number of parameters in increasing order,
with ties broken by the value of \verb{-2logLik} in decreasing order.}
}
\value{
A list of class \code{anova.mvreg}:
\itemize{
\item If only one model is provided:
\itemize{
\item \code{mu.tests}: A data frame with results of the ANOVA for the mean component, including:
\itemize{
\item \verb{-2logLik}: Negative two times the log-likelihood of each model.
\item \code{n.param}: Number of parameters in each model.
\item \code{AIC}: Akaike Information Criterion for each model.
\item \code{LRT}: Likelihood ratio test statistic.
\item \code{df}: Degrees of freedom for the test.
\item \verb{Pr(>Chi)}: P-value associated with the likelihood ratio test.
}
\item \code{s2.tests}: A similar data frame for the variance component.
\item \code{new.formula.mu}: A character vector of formulas used for the mean component.
\item \code{new.formula.s2}: A character vector of formulas used for the variance component.
\item \code{n_models}: Number of models compared.
}
\item If more than one model is provided:
\itemize{
\item \code{tests}: A data frame with results of the likelihood ratio tests, including:
\itemize{
\item \verb{-2logLik}: Negative two times the log-likelihood of each model.
\item \code{n.param}: Number of parameters in each model.
\item \code{AIC}: Akaike Information Criterion for each model.
\item \code{LRT}: Likelihood ratio test statistic.
\item \verb{Pr(>Chi)}: P-value associated with the likelihood ratio test.
}
\item \code{formulas.mu}: A character vector of formulas used for the mean component.
\item \code{formulas.s2}: A character vector of formulas used for the variance component.
\item \code{n_models}: Number of models compared.
}
}
}
\description{
This function performs an ANOVA-like analysis for comparing one or more \code{mvreg} models. It uses likelihood ratio tests (LRT)
to compare nested models, providing test statistics and p-values. If only one \code{mvreg} model is provided, the function
compares the model with reduced versions obtained by sequentially removing terms from the
mean and variance formulas.
}
\details{
The \code{anova.mvreg} function compares \code{mvreg} models using likelihood ratio tests
(LRT). If multiple models are provided, they are compared based on their likelihoods
and parameter counts. If only one model is provided, the function automatically generates
reduced models by sequentially dropping terms from the mean and variance formulas.
This allows for comparison of nested models and the identification of significant terms.

The function checks whether the models being compared are nested using the \code{are_models_nested()} function.
If non-nested models are provided, a warning is issued, as the likelihood ratio test is only valid for nested models.
Additionally, models with the same number of parameters cannot be meaningfully compared using the LRT.

If \code{order.models = TRUE}, models are ranked by the number of parameters (increasing order) and further sorted by
the \verb{-2logLik} value (decreasing order) if the number of parameters is equal. This ensures that models with the
same number of parameters are presented in order of likelihood.

The returned object contains detailed information about the comparison, including AIC values,
degrees of freedom, likelihood ratio statistics, and their associated p-values.
}
\note{
If two models have the same number of parameters and/or are nested, the likelihood ratio test is meaningless,
and the function will issue a warning. Negative LRT or degrees of freedom values are also flagged as errors.
}
\examples{
# Fit two models
mod1 <- mvreg(Sepal.Length ~ Species + Sepal.Width, data = iris)
mod2 <- mvreg(Sepal.Length ~ Species, ~1, data = iris)
anova(mod1, mod2)

# Compare a single model with its reduced forms
mod3 <- mvreg(Sepal.Length ~ Species, ~Sepal.Width, data = iris)
anova(mod3)
}
