% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_utilities.R
\name{get_reduced_formulas}
\alias{get_reduced_formulas}
\title{Generate formulas by dropping terms one by one}
\usage{
get_reduced_formulas(response, formula)
}
\arguments{
\item{response}{A character string representing the response variable to be used on the left-hand side of the formula.}

\item{formula}{A formula object from which to generate the reduced formulas.}
}
\value{
A list of formula objects, each representing a reduced version of the input formula by removing terms one by one.
}
\description{
This function generates a list of formulas by iteratively removing the last term
from the right-hand side of the provided formula, returning a series of reduced formulas.
}
\examples{
formula <- y ~ x1 * x2 * x3
response <- "y"
get_reduced_formulas(response, formula)
}
