% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{logLik.mvreg}
\alias{logLik.mvreg}
\title{logLik for mvreg}
\usage{
\method{logLik}{mvreg}(x)
}
\arguments{
\item{x}{A mvreg object.}
}
\value{
Loglikelihood value of a fitted mvreg model.
}
\description{
logLik for mvreg
}
