% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{logLik.mvreg}
\alias{logLik.mvreg}
\title{logLik for mvreg}
\usage{
\method{logLik}{mvreg}(object, ...)
}
\arguments{
\item{object}{A mvreg object.}

\item{...}{Additional optional arguments.}
}
\value{
Loglikelihood value of a fitted mvreg model.
}
\description{
logLik for mvreg
}
\examples{
mod <- mvreg(Sepal.Length ~ Species, data = iris)
logLik(mod)
}
