% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg.R
\name{mvreg}
\alias{mvreg}
\title{Heteroscedastic linear model with variance as a function of linear predictor.}
\usage{
mvreg(
  formula.mu,
  formula.s2 = NULL,
  data = NULL,
  tol = 1e-10,
  maxit = 100L,
  start.s2 = c("residuals", "gamma", "zero")
)
}
\arguments{
\item{formula.mu}{A two-sided formula describing the model for mean component.}

\item{formula.s2}{A one-sided formula describing the model for variance component.}

\item{data}{An optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{tol}{Positive value indicating what is the minimum difference between parameter estimates between two iterations to stop the algorithm.}

\item{maxit}{Integer value indicating the maximum number of iteration.}

\item{start.s2}{A character vector indicating how to select initial values for variance component parameters.}
}
\value{
An object of class mvreg
}
\description{
Heteroscedastic linear model with variance as a function of linear predictor.
}
