% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{mvreg_hessian}
\alias{mvreg_hessian}
\title{Full hessian of model log-likelihood}
\usage{
mvreg_hessian(y, x, z, b, t, type = c("observed", "expected"))
}
\arguments{
\item{y}{A numeric vector representing the response variable.}

\item{x}{A numeric matrix of explanatory variables for the mean component of the model.}

\item{z}{A numeric matrix of explanatory variables for the variance component of the model.}

\item{b}{A numeric vector of parameters for the mean component.}

\item{t}{A numeric vector of parameters for the variance component.}

\item{type}{A character string specifying whether to return the observed or expected Hessian matrix.
The default is "observed".}
}
\value{
A numeric matrix representing the full hessian of the log-likelihood with respect to the mean and variance component coefficients.
}
\description{
This function computes the full hessian matrix of the log-likelihood of a heteroskedastic linear model.
The hessian can be calculated as either the observed or expected matrix, depending on the \code{type} argument.
}
\details{
The full Hessian is constructed by combining the hessians of the mean and variance components,
as well as the mixed partial derivatives between them. The structure of the returned matrix is:
\deqn{\mathcal{H}(\boldsymbol{\beta}, \boldsymbol{\tau}) =
\begin{bmatrix}
\mathcal{H}_{\boldsymbol{\beta}}(\boldsymbol{\beta}, \boldsymbol{\tau}) & \mathcal{H}_{\boldsymbol{\beta} \boldsymbol{\tau}}(\boldsymbol{\beta}, \boldsymbol{\tau}) \\
\mathcal{H}_{\boldsymbol{\beta} \boldsymbol{\tau}}'(\boldsymbol{\beta}, \boldsymbol{\tau}) & \mathcal{H}_{\boldsymbol{\tau}}(\boldsymbol{\beta}, \boldsymbol{\tau})
\end{bmatrix}}
}
\examples{
# Fit a heteroskedastic linear model to the iris dataset
mvreg_mod <- mvreg(Sepal.Length ~ Species, data = iris) # same formula for mean and variance

# Extract model components
y <- mvreg_mod$y # Response variable
x <- mvreg_mod$x # Model matrix for the mean component
z <- mvreg_mod$z # Model matrix for the variance component
b <- coef(mvreg_mod, "mu") # Coefficients of the mean component
t <- coef(mvreg_mod, "s2") # Coefficients of the variance component

# Compute the full Hessian
mvreg_hessian(y, x, z, b, t, type = "observed")
mvreg_hessian(y, x, z, b, t, type = "expected")
}
