% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{mvreg_loglik}
\alias{mvreg_loglik}
\title{Log-likelihood of a Heteroskedastic Linear Model}
\usage{
mvreg_loglik(y, x, z, b, t)
}
\arguments{
\item{y}{A numeric vector representing the response variable.}

\item{x}{A numeric matrix of explanatory variables for the mean component of the model.}

\item{z}{A numeric matrix of explanatory variables for the variance component of the model.}

\item{b}{A numeric vector of parameters for the mean component.}

\item{t}{A numeric vector of parameters for the variance component.}
}
\value{
A numeric value representing the log-likelihood of the model for the given sample and parameter values.
}
\description{
This function computes the log-likelihood of a heteroskedastic linear model, where the variance component is modeled as a linear function of a set of explanatory variables, possibly different from those of the mean component.
}
\details{
The log-likelihood is computed under the assumption that the response variable follows a normal distribution with mean \eqn{\mathbf{x}_i' \boldsymbol{\beta}} and variance \eqn{\exp\left\{\mathbf{z}_i' \boldsymbol{\tau}\right\}}, where \eqn{\boldsymbol{\beta}} and \eqn{\boldsymbol{\tau}} are the parameters vectors for the mean and variance components, respectively.
The log-likelihood is computed as \deqn{\ell(\boldsymbol{\beta},\boldsymbol{\tau}) = -\dfrac{1}{2}\sum_{i=1}^n \left\{\log(2\pi) + \mathbf{z}_i'\boldsymbol{\tau} + \dfrac{\left(y_i - \mathbf{x}_i'\boldsymbol{\beta}\right)^2}{\exp\left\{\mathbf{z}_i'\boldsymbol{\tau}\right\}} \right\}}
}
\examples{
# Fit a heteroskedastic linear model to the iris dataset
mvreg_mod <- mvreg(Sepal.Length ~ Species, data = iris) # same formula for mean and variance

# Extract model components
y <- mvreg_mod$y # Response variable
x <- mvreg_mod$x # Model matrix for the mean component
z <- mvreg_mod$z # Model matrix for the variance component
b <- coef(mvreg_mod, "mu") # Coefficients of the mean component
t <- coef(mvreg_mod, "s2") # Coefficients of the variance component

# Compute log-likelihood
mvreg_loglik(y, x, z, b, t)
}
