% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_simul.R
\name{mvreg_simul}
\alias{mvreg_simul}
\title{Simulation study for mvreg models}
\usage{
mvreg_simul(
  x,
  z,
  b,
  t,
  n = 100,
  nsim = 100,
  sig.level = 0.05,
  seed = NULL,
  method = c("wls", "full_nr"),
  vcov.type = c("expected", "observed"),
  start.s2 = c("residuals", "gamma", "zero"),
  tol = 0.0000000001,
  maxit = 100
)
}
\arguments{
\item{x}{Design matrix for mean component.}

\item{z}{Design matrix for variance component.}

\item{b}{True values for parameters of mean component.}

\item{t}{True values for parameters of variance component.}

\item{n}{Sample size.}

\item{nsim}{Number of samples to simulate.}

\item{sig.level}{Significance level for Wald tests.}

\item{seed}{Seed of the simulation.}

\item{method}{Method chosen for estimation of parameters of mean component.}

\item{vcov.type}{A string to specify whether to use observed or expected Fisher information matrix in order to compute variance-covariance matrix of estimates.}

\item{start.s2}{A character vector indicating how to select initial values for variance component parameters.}

\item{tol}{Positive value indicating what is the minimum difference between parameter estimates between two iterations to stop the algorithm.}

\item{maxit}{Integer value indicating the maximum number of iteration.}
}
\value{
Results of a simulation study to check properties of estimators and statistical tests.
}
\description{
Simulation study for mvreg models
}
\examples{
n <- 100
x <- cbind(1, rnorm(n))
z <- cbind(1, x[,2], x[,2]^2)
b <- rnorm(ncol(x))
t <- rnorm(ncol(z))
mvreg_simul(x, z, b, t, n = n, nsim = 100, seed = 43)


}
