% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_to_lm.R
\name{mvreg_to_lm}
\alias{mvreg_to_lm}
\title{Convert mvreg model to lm}
\usage{
mvreg_to_lm(object)
}
\arguments{
\item{object}{A mvreg object}
}
\value{
A lm object with same formula as the formula.mu of the initial mvreg object
}
\description{
Convert mvreg model to lm
}
\examples{
mod.mvreg <- mvreg(dist ~ speed, data = cars)
mvreg_to_lm(mod.mvreg)
}
