% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{predict.mvreg}
\alias{predict.mvreg}
\title{Predict method for mvreg}
\usage{
\method{predict}{mvreg}(
  object,
  type = c("all", "mu", "log.s2", "s2"),
  newdata,
  se.fit = F,
  interval = F,
  level = 0.95
)
}
\arguments{
\item{object}{A mvreg object.}

\item{type}{A character vector specifying the component for which to return predicted values.}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{se.fit}{A logical value indicating if standard errors are to be returned.}

\item{interval}{A logical value indicating if confidence intervals are to be returned.}

\item{level}{Confidence level for confidence intervals.}
}
\value{
Predicted values for the different components of mvreg model. If specified, standard error estimates and confidence interval are returned.
}
\description{
Predict method for mvreg
}
\examples{
mvreg_mod1 <- mvreg(Sepal.Length ~ Species, ~Sepal.Width, data = iris) # different formulas

# predict without newdata
predict(mvreg_mod1)
predict(mvreg_mod1, type = "mu")
predict(mvreg_mod1, type = "log.s2", se.fit = TRUE)
predict(mvreg_mod1, type = "s2", interval = TRUE)
predict(mvreg_mod1, se.fit = TRUE, interval = TRUE, level = 0.99)

# predict with newdata
unique(c(mvreg_mod1$colx, mvreg_mod1$colz)) # getting names of explanatory variables

newdata <- data.frame(
  Species = levels(iris$Species),
  Sepal.Width = c(min(iris$Sepal.Width), mean(iris$Sepal.Width), max(iris$Sepal.Width))
)

newdata <- expand.grid(newdata)

predict(mvreg_mod1, newdata = newdata, se.fit = TRUE, interval = TRUE)
}
