% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{print.mvreg}
\alias{print.mvreg}
\title{Print method for mvreg}
\usage{
\method{print}{mvreg}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{A mvreg object.}

\item{digits}{Minimal number of significant digits, see print.default.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Print informations about the model.
}
\description{
Print method for mvreg
}
