% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{print.summary.mvreg}
\alias{print.summary.mvreg}
\title{Print method for summary.mvreg}
\usage{
\method{print}{summary.mvreg}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{A mvreg object.}

\item{digits}{Minimal number of significant digits, see print.default.}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Print summary for mvreg.
}
\description{
Print method for summary.mvreg
}
\examples{
mod <- mvreg(Sepal.Length ~ Species, data = iris)
s <- summary(mod)
print(s)
}
