% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{simulate.mvreg}
\alias{simulate.mvreg}
\title{Simulate Responses from a mvreg object}
\usage{
\method{simulate}{mvreg}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{A mvreg object.}

\item{nsim}{Number of response vectors to simulate}

\item{seed}{NULL or an integer that will be used in a call to \code{set.seed} before simulating the response vectors. If set, the value is saved as the "seed" attribute of the returned value. The default, NULL will not change the random generator state, and return .Random.seed as the "seed" attribute.}

\item{...}{Additional optional arguments.}
}
\value{
A data.frame which columns are simulated response vectors.
}
\description{
Simulate Responses from a mvreg object
}
\examples{
mvreg.mod <- mvreg(Sepal.Length ~ Species, data = iris)
simulate(mvreg.mod, nsim = 100, seed = 43)

}
