% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{summary.mvreg}
\alias{summary.mvreg}
\title{Summary method for mvreg}
\usage{
\method{summary}{mvreg}(object, ...)
}
\arguments{
\item{object}{A mvreg object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A summary.mvreg object.
}
\description{
Summary method for mvreg
}
\examples{
mvreg_mod <- mvreg(Sepal.Length ~ Species, data = iris) # same formula for mean and variance
mvreg_mod1 <- mvreg(Sepal.Length ~ Species, ~ Sepal.Width, data = iris) # different formulas
summary(mvreg_mod1)
summary(mvreg_mod)
}
