% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{update.mvreg}
\alias{update.mvreg}
\title{Update and Re-fit a mvreg model}
\usage{
\method{update}{mvreg}(object, new.formula.mu, new.formula.s2, ..., evaluate = TRUE)
}
\arguments{
\item{object}{A mvreg object.}

\item{new.formula.mu}{Changes to formula.mu – see update.formula for details.}

\item{new.formula.s2}{Changes to formula.s2 – see update.formula for details.}

\item{...}{Additional arguments to the call, or arguments with changed values. Use name = NULL to remove the argument name.}

\item{evaluate}{If true evaluate the new call else return the call.}
}
\value{
A new mvreg object with the updated formulas or the call
}
\description{
Update and Re-fit a mvreg model
}
\examples{
mod <- mvreg(Sepal.Length ~ Species, data = iris)

# changing formulas
update(mod, new.formula.mu = . ~ . + Sepal.Width)
update(mod, new.formula.mu = . ~ . + Sepal.Width, new.formula.s2 = . ~ . + Petal.Length)

# changing method, vcov.type, start.s2
update(mod, method = "full_nr", vcov.type = "observed", start.s2 = "gamma")
}
