% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvreg_methods.R
\name{vcov.mvreg}
\alias{vcov.mvreg}
\title{Variance-Covariance matrix for mvreg}
\usage{
\method{vcov}{mvreg}(object, partition = c("all", "mu", "s2"), ...)
}
\arguments{
\item{object}{A mvreg object.}

\item{partition}{A character vector that specifies which partition of the matrix to return.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Specified partition of Variance-Covariance matrix of model's parameters.
}
\description{
Variance-Covariance matrix for mvreg
}
\examples{
mvreg_mod <- mvreg(Sepal.Length ~ Species, ~Species, data = iris)
summary(mvreg_mod)
vcov(mvreg_mod)
vcov(mvreg_mod, partition = "mu")
}
